refactor: divide resolution and fetching
