+++ src/resolve/git.ts
@@ -1,9 +1,8 @@
 import spawn = require('cross-spawn')
 import {PackageSpec, ResolveOptions, ResolveResult} from '.'
 import {delimiter} from './createPkgId'
-import createDebug from '../debug'
-const debug = createDebug('pnpm:git')
 
+// TODO: resolve ref to commit
 export default async function resolveGithub (parsedSpec: PackageSpec, opts: ResolveOptions): Promise<ResolveResult> {
   const parts = parsedSpec.spec.split('#')
   const repo = parts[0]
@@ -13,38 +12,7 @@ export default async function resolveGithub (parsedSpec: PackageSpec, opts: Reso
       .replace(/^.*:\/\/(git@)?/, '')
       .replace(/:/g, delimiter)
       .replace(/\.git$/, ''),
-    fetch: (target: string) => {
-      return clone(repo, ref, target)
-    }
+    repo,
+    ref,
   }
-}
-
-function prefixGitArgs (): string[] {
-  return process.platform === 'win32' ? ['-c', 'core.longpaths=true'] : []
-}
-
-function spawnGit (args: string[]) {
-  debug(`executing git with args ${args}`)
-  const fullArgs = prefixGitArgs().concat(args || [])
-  return spawn('git', fullArgs)
-}
-
-/**
- * clone a git repository.
- */
-export async function clone (repo: string, ref: string, dest: string) {
-  await new Promise((resolve, reject) => {
-    const args = ['clone', '-b', ref, repo, dest, '--single-branch']
-    debug(`cloning git repository from ${repo}`)
-    const git = spawnGit(args)
-    let errMsg = ''
-    git.stderr.on('data', (data: string) => errMsg += data)
-    git.on('close', (code: number) => (code ? errorHandler() : resolve()))
-
-    function errorHandler () {
-      debug(`failed to clone repository from ${repo}`)
-      reject(new Error(`failed to clone repository from ${repo}
-        ${errMsg}`))
-    }
-  })
 }
